+++ test/hcp/account_test.rb
@@ -33,6 +33,6 @@ class AccountTestCase < Minitest::Test
   end
 
   def test_refuses_what_a_key_does_not_open
-    assert_raises(NotImplementedError) { account.jobs }
+    assert_raises(NotImplementedError) { account.visits }
   end
 end